+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.6.0](https://github.com/crypyto-panel/etd-local-admin-node/compare/v1.5.3...v1.6.0) (2021-11-21)
+
+
+### Features
+
+* Add action to pull requests ([db4678d](https://github.com/crypyto-panel/etd-local-admin-node/commit/db4678de0c7ea9384f0381b9c9576dcee92af6a5))
+* Add docker data in status plugin ([f7c28e9](https://github.com/crypyto-panel/etd-local-admin-node/commit/f7c28e959ea8863886ffa347898d6e8f09f63952))
+* CI/CD improvements ([e760104](https://github.com/crypyto-panel/etd-local-admin-node/commit/e7601042f8ce67f696722465a095899181c1fa07))
+
+
+### Bug Fixes
+
+* Package.json ([c23b762](https://github.com/crypyto-panel/etd-local-admin-node/commit/c23b7628a9df975385549939be6a571926ca3a82))
+
 ### [1.5.3](https://github.com/crypyto-panel/etd-local-admin-node/compare/v1.5.2...v1.5.3) (2021-10-15)
 
 ### [1.5.2](https://github.com/crypyto-panel/etd-local-admin-node/compare/v1.5.1...v1.5.2) (2021-10-15)